Connect Four
Properly installed the numpy and pygame file by checking to see if the pip file was installed. I upgraded the pip to pip 22.1.2. I installed the numpy and pygame file by "pip insall numpy" and "pip install pygame". The install was successful and the code was officially able to run